#181 To make it easier to make adjustments to the lower panel of the gui I have moved all of the lower panel components into its own file called viewLowerPanel.  I have tested the application several times to make sure nothing was broken.  Everything is working correctly.
